simulation/requirements.txt
requests==2.32.5
Pillow==12.1.0